Add photo prompt and instructions to back photo flow
